Make `tracing.trace` idempotent, and allow it to be applied to a function that was already decorated with `@executing.make_executable`.
The motivation for allowing `tracing.trace` to be applied to a function that was already decorated with `@executing.make_executable` is to make it easier to implement (among other things) a generic evaluation framework that can take as input a prompting strategy implemented as an arbitrary function (which may or may not already be traced, and may or may not already be decorated with `@executing.make_executable`), and then apply to it an extra `tracing.trace` layer, just in case, to ensure that at minimum the outermost inputs/outputs of the prompting strategies are traced.

Note that it is still recommended, where possible, to decorate a function with `@tracing.trace` first, and then with `@executing.make_executable` outside of that, as there are some limitations in behavior when the decorators are reversed. In particular, tracing does not work reliably when `executing.run` or `executing.stream_updates`, etc. is performed directly on a function for which the outermost decorator was `@tracing.trace`. For this reason, in cases such as the motivating example above, where one wants to apply an extra layer of `tracing.trace` "just in case", it is best to then wrap that in one more function layer that is decorated with `@executing.make_executable` in the usual way. (See the test cases added in this CL for details.)

PiperOrigin-RevId: 631787870
Change-Id: Ibe83c6191cd5fbbd126367f27b40d057c8c36132